% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadST-main-functions.R
\name{ICG_distance}
\alias{ICG_distance}
\title{Find cell-cell interaction distance (given ICG is identified)}
\usage{
ICG_distance(x, ICG.summary, k)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} class.}

\item{ICG.summary}{ICG summary.}

\item{k}{K nearest neighbor.}
}
\value{
Interacting distance.
}
\description{
Find cell-cell interaction distance (given ICG is identified)
}
