% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadST-main-functions.R
\name{identify_ICGs}
\alias{identify_ICGs}
\title{Identify cell-cell interaction changes genes (ICGs)}
\usage{
identify_ICGs(pMatrix, fdr = 0.1)
}
\arguments{
\item{pMatrix}{A matrix of p-values in the genes (rows) by quantile levels (columns) format.}

\item{fdr}{Default= 0.1. The nominal false discovery rate to claim significance.}
}
\value{
a list of inference results
\item{summary.table}{A summary of the results, including the quantile index and level of the
cell-cell interaction, and the No. of ICGs significant at this quantile level.}
\item{data.table}{A table of gene-level results, including gene name, p-value at the specified quantile
level, the empirical FDR, and whether it's a ICG or not.}
}
\description{
Identify cell-cell interaction changes genes (ICGs)
}
