% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadST-main-functions.R
\name{test_QuadST_model}
\alias{test_QuadST_model}
\title{Test anchor-neighbor distance-expression association
 at a set of highest and lowest quantiles symmetric around median}
\usage{
test_QuadST_model(x, datatype, cov = NULL, tau, parallel = F)
}
\arguments{
\item{x}{A \code{SingleCellExperiment} class.}

\item{datatype}{A column name of \code{assays(object)} that stores anchor cells' gene expression levels.}

\item{cov}{Column names of \code{colData(object)} that needs to be adjusted as covariates.}

\item{tau}{A set of highest and lowest quantiles symmetric around median.}

\item{parallel}{Default=False. If parallel computing is activated.}
}
\value{
A matrix of quantile regression p-values in the format of genes (rows) by quantile levels (columns).
}
\description{
Test anchor-neighbor distance-expression association
 at a set of highest and lowest quantiles symmetric around median
}
