% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadST-helper-functions.R
\name{transform_count_to_normal}
\alias{transform_count_to_normal}
\title{Transform cell-specific bias adjusted counts to a normally distribution with the minimum value at zero}
\usage{
transform_count_to_normal(x)
}
\arguments{
\item{x}{An expression matrix.}
}
\value{
A transformed expression matrix.
}
\description{
Transform cell-specific bias adjusted counts to a normally distribution with the minimum value at zero
}
\seealso{
[test_QuadST_model()] to see how this function is used.
}
